MueLu: Merge SaPFactory_kokkos and SaPFactory
