+++ packages/muelu/src/Interface/MueLu_FactoryFactory_decl.hpp
@@ -163,8 +163,8 @@
 #include "MueLu_GeometricInterpolationPFactory_kokkos.hpp"
 #ifdef HAVE_MUELU_DEPRECATED_CODE
 #include "MueLu_NullspaceFactory_kokkos.hpp"
-#endif
 #include "MueLu_SaPFactory_kokkos.hpp"
+#endif
 #include "MueLu_SemiCoarsenPFactory_kokkos.hpp"
 #include "MueLu_StructuredAggregationFactory_kokkos.hpp"
 #include "MueLu_TentativePFactory_kokkos.hpp"
@@ -308,8 +308,8 @@ class FactoryFactory : public BaseClass {
     if (factoryName == "GeometricInterpolationPFactory_kokkos") return Build2<GeometricInterpolationPFactory_kokkos>(paramList, factoryMapIn, factoryManagersIn);
 #ifdef HAVE_MUELU_DEPRECATED_CODE
     if (factoryName == "NullspaceFactory_kokkos") return Build2<NullspaceFactory_kokkos>(paramList, factoryMapIn, factoryManagersIn);
-#endif
     if (factoryName == "SaPFactory_kokkos") return Build2<SaPFactory_kokkos>(paramList, factoryMapIn, factoryManagersIn);
+#endif
     if (factoryName == "SemiCoarsenPFactory_kokkos") return Build2<SemiCoarsenPFactory_kokkos>(paramList, factoryMapIn, factoryManagersIn);
     if (factoryName == "StructuredAggregationFactory_kokkos") return Build2<StructuredAggregationFactory_kokkos>(paramList, factoryMapIn, factoryManagersIn);
     if (factoryName == "TentativePFactory_kokkos") return Build2<TentativePFactory_kokkos>(paramList, factoryMapIn, factoryManagersIn);
