+++ packages/muelu/src/Interface/MueLu_MLParameterListInterpreter_def.hpp
@@ -85,7 +85,6 @@
 
 #include "MueLu_CoalesceDropFactory_kokkos.hpp"
 // #include "MueLu_CoordinatesTransferFactory_kokkos.hpp"
-#include "MueLu_SaPFactory_kokkos.hpp"
 #include "MueLu_TentativePFactory_kokkos.hpp"
 
 #if defined(HAVE_MUELU_ISORROPIA) && defined(HAVE_MPI)
@@ -290,11 +289,7 @@ void MLParameterListInterpreter<Scalar, LocalOrdinal, GlobalOrdinal, Node>::SetP
     RFact = rcp(new TransPFactory());
   } else if (agg_damping != 0.0 && bEnergyMinimization == false) {
     // smoothed aggregation (SA-AMG)
-    RCP<Factory> SaPFact;
-    if (useKokkosRefactor)
-      SaPFact = rcp(new SaPFactory_kokkos());
-    else
-      SaPFact = rcp(new SaPFactory());
+    RCP<Factory> SaPFact = rcp(new SaPFactory());
     SaPFact->SetParameter("sa: damping factor", ParameterEntry(agg_damping));
     PFact = SaPFact;
     RFact = rcp(new TransPFactory());
