+++ packages/muelu/src/MueCentral/MueLu_FactoryManager_decl.hpp
@@ -76,7 +76,6 @@
 #include "MueLu_InterfaceAggregationFactory_fwd.hpp"
 
 #include "MueLu_CoalesceDropFactory_kokkos_fwd.hpp"
-#include "MueLu_SaPFactory_kokkos_fwd.hpp"
 #include "MueLu_TentativePFactory_kokkos_fwd.hpp"
 
 namespace MueLu {
