+++ packages/muelu/src/Operators/MueLu_RefMaxwell_decl.hpp
@@ -67,7 +67,6 @@
 
 #include "MueLu_CoalesceDropFactory_kokkos_fwd.hpp"
 #include "MueLu_TentativePFactory_kokkos_fwd.hpp"
-#include "MueLu_SaPFactory_kokkos_fwd.hpp"
 
 #include "MueLu_ZoltanInterface_fwd.hpp"
 #include "MueLu_Zoltan2Interface_fwd.hpp"
