+++ packages/muelu/src/Transfers/Smoothed-Aggregation/MueLu_SaPFactory_def.hpp
@@ -46,12 +46,11 @@
 #ifndef MUELU_SAPFACTORY_DEF_HPP
 #define MUELU_SAPFACTORY_DEF_HPP
 
-#include <Xpetra_Matrix.hpp>
-#include <Xpetra_IteratorOps.hpp>  // containing routines to generate Jacobi iterator
-#include <Xpetra_IO.hpp>
-#include <sstream>
-
+#include "Kokkos_ArithTraits.hpp"
 #include "MueLu_SaPFactory_decl.hpp"
+
+#include <Xpetra_Matrix.hpp>
+#include <Xpetra_IteratorOps.hpp>
 
 #include "MueLu_FactoryManagerBase.hpp"
 #include "MueLu_Level.hpp"
@@ -60,6 +59,8 @@
 #include "MueLu_PerfUtils.hpp"
 #include "MueLu_TentativePFactory.hpp"
 #include "MueLu_Utilities.hpp"
+
+#include <sstream>
 
 namespace MueLu {
 
@@ -79,6 +80,7 @@ RCP<const ParameterList> SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::
   SET_VALID_ENTRY("sa: rowsumabs diagonal replacement value");
   SET_VALID_ENTRY("sa: rowsumabs replace single entry row with zero");
   SET_VALID_ENTRY("sa: rowsumabs use automatic diagonal tolerance");
+  SET_VALID_ENTRY("use kokkos refactor");
 #undef SET_VALID_ENTRY
 
   validParamList->set<RCP<const FactoryBase> >("A", Teuchos::null, "Generating factory of the matrix A used during the prolongator smoothing process");
@@ -93,7 +95,7 @@ RCP<const ParameterList> SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::
 }
 
 template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-void SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::DeclareInput(Level &fineLevel, Level &coarseLevel) const {
+void SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::DeclareInput(Level& fineLevel, Level& coarseLevel) const {
   Input(fineLevel, "A");
 
   // Get default tentative prolongator factory
@@ -102,16 +104,16 @@ void SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::DeclareInput(Level &
   if (initialPFact == Teuchos::null) {
     initialPFact = coarseLevel.GetFactoryManager()->GetFactory("Ptent");
   }
-  coarseLevel.DeclareInput("P", initialPFact.get(), this);  // --
+  coarseLevel.DeclareInput("P", initialPFact.get(), this);
 }
 
 template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-void SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::Build(Level &fineLevel, Level &coarseLevel) const {
+void SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::Build(Level& fineLevel, Level& coarseLevel) const {
   return BuildP(fineLevel, coarseLevel);
 }
 
 template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-void SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::BuildP(Level &fineLevel, Level &coarseLevel) const {
+void SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::BuildP(Level& fineLevel, Level& coarseLevel) const {
   FactoryMonitor m(*this, "Prolongator smoothing", coarseLevel);
 
   std::string levelIDs = toString(coarseLevel.GetLevelID());
@@ -128,7 +130,7 @@ void SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::BuildP(Level &fineLe
   if (initialPFact == Teuchos::null) {
     initialPFact = coarseLevel.GetFactoryManager()->GetFactory("Ptent");
   }
-  const ParameterList &pL = GetParameterList();
+  const ParameterList& pL = GetParameterList();
 
   // Level Get
   RCP<Matrix> A     = Get<RCP<Matrix> >(fineLevel, "A");
@@ -144,7 +146,7 @@ void SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::BuildP(Level &fineLe
 
   if (restrictionMode_) {
     SubFactoryMonitor m2(*this, "Transpose A", coarseLevel);
-    A = Utilities::Transpose(*A, true);  // build transpose of A explicitely
+    A = Utilities::Transpose(*A, true);  // build transpose of A explicitly
   }
 
   // Build final prolongator
@@ -155,7 +157,6 @@ void SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::BuildP(Level &fineLe
     APparams = rcp(new ParameterList(pL.sublist("matrixmatrix: kernel params")));
   else
     APparams = rcp(new ParameterList);
-
   if (coarseLevel.IsAvailable("AP reuse data", this)) {
     GetOStream(static_cast<MsgType>(Runtime0 | Test)) << "Reusing previous AP data" << std::endl;
 
@@ -168,19 +169,18 @@ void SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::BuildP(Level &fineLe
   APparams->set("compute global constants: temporaries", APparams->get("compute global constants: temporaries", false));
   APparams->set("compute global constants", APparams->get("compute global constants", false));
 
-  const SC dampingFactor        = as<SC>(pL.get<double>("sa: damping factor"));
-  const LO maxEigenIterations   = as<LO>(pL.get<int>("sa: eigenvalue estimate num iterations"));
-  const bool estimateMaxEigen   = pL.get<bool>("sa: calculate eigenvalue estimate");
-  const bool useAbsValueRowSum  = pL.get<bool>("sa: use rowsumabs diagonal scaling");
-  const bool doQRStep           = pL.get<bool>("tentative: calculate qr");
-  const bool enforceConstraints = pL.get<bool>("sa: enforce constraints");
-  const MT userDefinedMaxEigen  = as<MT>(pL.get<double>("sa: max eigenvalue"));
-  typedef typename Teuchos::ScalarTraits<Scalar>::magnitudeType Magnitude;
-  double dTol                                  = pL.get<double>("sa: rowsumabs diagonal replacement tolerance");
-  const Magnitude diagonalReplacementTolerance = (dTol == as<double>(-1) ? Teuchos::ScalarTraits<Scalar>::eps() * 100 : as<Magnitude>(pL.get<double>("sa: rowsumabs diagonal replacement tolerance")));
-  const SC diagonalReplacementValue            = as<SC>(pL.get<double>("sa: rowsumabs diagonal replacement value"));
-  const bool replaceSingleEntryRowWithZero     = pL.get<bool>("sa: rowsumabs replace single entry row with zero");
-  const bool useAutomaticDiagTol               = pL.get<bool>("sa: rowsumabs use automatic diagonal tolerance");
+  const SC dampingFactor                   = as<SC>(pL.get<double>("sa: damping factor"));
+  const LO maxEigenIterations              = as<LO>(pL.get<int>("sa: eigenvalue estimate num iterations"));
+  const bool estimateMaxEigen              = pL.get<bool>("sa: calculate eigenvalue estimate");
+  const bool useAbsValueRowSum             = pL.get<bool>("sa: use rowsumabs diagonal scaling");
+  const bool doQRStep                      = pL.get<bool>("tentative: calculate qr");
+  const bool enforceConstraints            = pL.get<bool>("sa: enforce constraints");
+  const MT userDefinedMaxEigen             = as<MT>(pL.get<double>("sa: max eigenvalue"));
+  double dTol                              = pL.get<double>("sa: rowsumabs diagonal replacement tolerance");
+  const MT diagonalReplacementTolerance    = (dTol == as<double>(-1) ? Teuchos::ScalarTraits<MT>::eps() * 100 : as<MT>(pL.get<double>("sa: rowsumabs diagonal replacement tolerance")));
+  const SC diagonalReplacementValue        = as<SC>(pL.get<double>("sa: rowsumabs diagonal replacement value"));
+  const bool replaceSingleEntryRowWithZero = pL.get<bool>("sa: rowsumabs replace single entry row with zero");
+  const bool useAutomaticDiagTol           = pL.get<bool>("sa: rowsumabs use automatic diagonal tolerance");
 
   // Sanity checking
   TEUCHOS_TEST_FOR_EXCEPTION(doQRStep && enforceConstraints, Exceptions::RuntimeError,
@@ -235,14 +235,23 @@ void SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::BuildP(Level &fineLe
       SC omega = dampingFactor / lambdaMax;
       TEUCHOS_TEST_FOR_EXCEPTION(!std::isfinite(Teuchos::ScalarTraits<SC>::magnitude(omega)), Exceptions::RuntimeError, "Prolongator damping factor needs to be finite.");
 
-      // finalP = Ptent + (I - \omega D^{-1}A) Ptent
-      finalP = Xpetra::IteratorOps<Scalar, LocalOrdinal, GlobalOrdinal, Node>::Jacobi(omega, *invDiag, *A, *Ptent, finalP,
-                                                                                      GetOStream(Statistics2), std::string("MueLu::SaP-") + levelIDs, APparams);
-      if (enforceConstraints) {
-        if (A->GetFixedBlockSize() == 1)
-          optimalSatisfyPConstraintsForScalarPDEs(finalP);
-        else
-          SatisfyPConstraints(A, finalP);
+      {
+        SubFactoryMonitor m3(*this, "Xpetra::IteratorOps::Jacobi", coarseLevel);
+        // finalP = Ptent + (I - \omega D^{-1}A) Ptent
+        finalP = Xpetra::IteratorOps<Scalar, LocalOrdinal, GlobalOrdinal, Node>::Jacobi(omega, *invDiag, *A, *Ptent, finalP, GetOStream(Statistics2), std::string("MueLu::SaP-") + toString(coarseLevel.GetLevelID()), APparams);
+        if (enforceConstraints) {
+          if (!pL.get<bool>("use kokkos refactor")) {
+            if (A->GetFixedBlockSize() == 1)
+              optimalSatisfyPConstraintsForScalarPDEsNonKokkos(finalP);
+            else
+              SatisfyPConstraintsNonKokkos(A, finalP);
+          } else {
+            // if (A->GetFixedBlockSize() == 1)
+            //   optimalSatisfyPConstraintsForScalarPDEs(finalP);
+            // else
+            SatisfyPConstraints(A, finalP);
+          }
+        }
       }
     }
 
@@ -251,8 +260,9 @@ void SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::BuildP(Level &fineLe
   }
 
   // Level Set
+  RCP<Matrix> R;
   if (!restrictionMode_) {
-    // The factory is in prolongation mode
+    // prolongation factory is in prolongation mode
     if (!finalP.is_null()) {
       std::ostringstream oss;
       oss << "P_" << coarseLevel.GetLevelID();
@@ -267,16 +277,8 @@ void SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::BuildP(Level &fineLe
     if (Ptent->IsView("stridedMaps"))
       finalP->CreateView("stridedMaps", Ptent);
 
-    if (IsPrint(Statistics2)) {
-      RCP<ParameterList> params = rcp(new ParameterList());
-      params->set("printLoadBalancingInfo", true);
-      params->set("printCommInfo", true);
-      GetOStream(Statistics2) << PerfUtils::PrintMatrixInfo(*finalP, "P", params);
-    }
-
   } else {
-    // The factory is in restriction mode
-    RCP<Matrix> R;
+    // prolongation factory is in restriction mode
     {
       SubFactoryMonitor m2(*this, "Transpose P", coarseLevel);
       R = Utilities::Transpose(*finalP, true);
@@ -292,13 +294,13 @@ void SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::BuildP(Level &fineLe
     // NOTE: EXPERIMENTAL
     if (Ptent->IsView("stridedMaps"))
       R->CreateView("stridedMaps", Ptent, true /*transposeA*/);
+  }
 
-    if (IsPrint(Statistics2)) {
-      RCP<ParameterList> params = rcp(new ParameterList());
-      params->set("printLoadBalancingInfo", true);
-      params->set("printCommInfo", true);
-      GetOStream(Statistics2) << PerfUtils::PrintMatrixInfo(*R, "R", params);
-    }
+  if (IsPrint(Statistics2)) {
+    RCP<ParameterList> params = rcp(new ParameterList());
+    params->set("printLoadBalancingInfo", true);
+    params->set("printCommInfo", true);
+    GetOStream(Statistics2) << PerfUtils::PrintMatrixInfo((!restrictionMode_ ? *finalP : *R), (!restrictionMode_ ? "P" : "R"), params);
   }
 
 }  // Build()
@@ -320,7 +322,7 @@ void SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::BuildP(Level &fineLe
 // the constraints.
 
 template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-void SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::SatisfyPConstraints(const RCP<Matrix> A, RCP<Matrix> &P) const {
+void SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::SatisfyPConstraintsNonKokkos(const RCP<Matrix> A, RCP<Matrix>& P) const {
   const Scalar zero = Teuchos::ScalarTraits<Scalar>::zero();
   const Scalar one  = Teuchos::ScalarTraits<Scalar>::one();
   LO nPDEs          = A->GetFixedBlockSize();
@@ -339,7 +341,7 @@ void SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::SatisfyPConstraints(
     size_t nnz = indices.size();
     if (nnz == 0) continue;
 
-    vals = ArrayView<Scalar>(const_cast<SC *>(vals1.getRawPtr()), nnz);
+    vals = ArrayView<Scalar>(const_cast<SC*>(vals1.getRawPtr()), nnz);
 
     bool checkRow = true;
 
@@ -395,7 +397,7 @@ void SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::SatisfyPConstraints(
 }  // SatsifyPConstraints()
 
 template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-void SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::optimalSatisfyPConstraintsForScalarPDEs(RCP<Matrix> &P) const {
+void SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::optimalSatisfyPConstraintsForScalarPDEsNonKokkos(RCP<Matrix>& P) const {
   const Scalar zero = Teuchos::ScalarTraits<Scalar>::zero();
   const Scalar one  = Teuchos::ScalarTraits<Scalar>::one();
 
@@ -410,7 +412,7 @@ void SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::optimalSatisfyPConst
     P->getLocalRowView((LocalOrdinal)i, indices, vals1);
     size_t nnz = indices.size();
     if (nnz != 0) {
-      vals              = ArrayView<Scalar>(const_cast<SC *>(vals1.getRawPtr()), nnz);
+      vals              = ArrayView<Scalar>(const_cast<SC*>(vals1.getRawPtr()), nnz);
       Scalar rsumTarget = zero;
       for (size_t j = 0; j < nnz; j++) rsumTarget += vals[j];
 
@@ -429,7 +431,7 @@ void SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::optimalSatisfyPConst
 }  // SatsifyPConstraints()
 
 template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-bool SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::constrainRow(Scalar *orig, LocalOrdinal nEntries, Scalar leftBound, Scalar rghtBound, Scalar rsumTarget, Scalar *fixedUnsorted, Scalar *scalarData) const {
+bool SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::constrainRow(Scalar* orig, LocalOrdinal nEntries, Scalar leftBound, Scalar rghtBound, Scalar rsumTarget, Scalar* fixedUnsorted, Scalar* scalarData) const {
   /*
      Input
        orig          data that should be adjusted to satisfy bound constraints and
@@ -551,7 +553,7 @@ bool SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::constrainRow(Scalar 
   Scalar rowSumDeviation, temp, *fixedSorted, delta;
   Scalar closestToLeftBoundDist, closestToRghtBoundDist;
   LocalOrdinal closestToLeftBound, closestToRghtBound;
-  LocalOrdinal *inds;
+  LocalOrdinal* inds;
   bool flipped;
 
   notFlippedLeftBound = leftBound;
@@ -576,7 +578,7 @@ bool SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::constrainRow(Scalar 
 
   origSorted  = &scalarData[0];
   fixedSorted = &(scalarData[nEntries]);
-  inds        = (LocalOrdinal *)&(scalarData[2 * nEntries]);
+  inds        = (LocalOrdinal*)&(scalarData[2 * nEntries]);
 
   for (LocalOrdinal i = 0; i < nEntries; i++) inds[i] = i;
   for (LocalOrdinal i = 0; i < nEntries; i++) origSorted[i] = orig[i]; /* orig no longer used */
@@ -717,6 +719,320 @@ bool SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::constrainRow(Scalar 
 
   return hasFeasibleSol;
 }
+
+template <typename local_matrix_type>
+struct constraintKernel {
+  using Scalar      = typename local_matrix_type::non_const_value_type;
+  using SC          = Scalar;
+  using LO          = typename local_matrix_type::non_const_ordinal_type;
+  using Device      = typename local_matrix_type::device_type;
+  using KAT         = Kokkos::ArithTraits<SC>;
+  const Scalar zero = KAT::zero();
+  const Scalar one  = KAT::one();
+  LO nPDEs;
+  local_matrix_type localP;
+  Kokkos::View<SC**, Device> ConstraintViolationSum;
+  Kokkos::View<SC**, Device> Rsum;
+  Kokkos::View<size_t**, Device> nPositive;
+
+  constraintKernel(LO nPDEs_, local_matrix_type localP_)
+    : nPDEs(nPDEs_)
+    , localP(localP_) {
+    ConstraintViolationSum = Kokkos::View<SC**, Device>("ConstraintViolationSum", localP_.numRows(), nPDEs);
+    Rsum                   = Kokkos::View<SC**, Device>("Rsum", localP_.numRows(), nPDEs);
+    nPositive              = Kokkos::View<size_t**, Device>("nPositive", localP_.numRows(), nPDEs);
+  }
+
+  KOKKOS_INLINE_FUNCTION
+  void operator()(const size_t rowIdx) const {
+    auto rowPtr = localP.graph.row_map;
+    auto values = localP.values;
+
+    bool checkRow = true;
+
+    if (rowPtr(rowIdx + 1) == rowPtr(rowIdx)) checkRow = false;
+
+    while (checkRow) {
+      // check constraints and compute the row sum
+
+      for (auto entryIdx = rowPtr(rowIdx); entryIdx < rowPtr(rowIdx + 1); entryIdx++) {
+        Rsum(rowIdx, entryIdx % nPDEs) += values(entryIdx);
+        if (KAT::real(values(entryIdx)) < KAT::real(zero)) {
+          ConstraintViolationSum(rowIdx, entryIdx % nPDEs) += values(entryIdx);
+          values(entryIdx) = zero;
+        } else {
+          if (KAT::real(values(entryIdx)) != KAT::real(zero))
+            nPositive(rowIdx, entryIdx % nPDEs) = nPositive(rowIdx, entryIdx % nPDEs) + 1;
+
+          if (KAT::real(values(entryIdx)) > KAT::real(1.00001)) {
+            ConstraintViolationSum(rowIdx, entryIdx % nPDEs) += (values(entryIdx) - one);
+            values(entryIdx) = one;
+          }
+        }
+      }
+
+      checkRow = false;
+
+      // take into account any row sum that violates the contraints
+
+      for (size_t k = 0; k < (size_t)nPDEs; k++) {
+        if (KAT::real(Rsum(rowIdx, k)) < KAT::magnitude(zero)) {
+          ConstraintViolationSum(rowIdx, k) = ConstraintViolationSum(rowIdx, k) - Rsum(rowIdx, k);  // rstumin
+        } else if (KAT::real(Rsum(rowIdx, k)) > KAT::magnitude(1.00001)) {
+          ConstraintViolationSum(rowIdx, k) = ConstraintViolationSum(rowIdx, k) + (one - Rsum(rowIdx, k));  // rstumin
+        }
+      }
+
+      // check if row need modification
+      for (size_t k = 0; k < (size_t)nPDEs; k++) {
+        if (KAT::magnitude(ConstraintViolationSum(rowIdx, k)) != KAT::magnitude(zero))
+          checkRow = true;
+      }
+      // modify row
+      if (checkRow) {
+        for (auto entryIdx = rowPtr(rowIdx); entryIdx < rowPtr(rowIdx + 1); entryIdx++) {
+          if (KAT::real(values(entryIdx)) > KAT::real(zero)) {
+            values(entryIdx) = values(entryIdx) +
+                               (ConstraintViolationSum(rowIdx, entryIdx % nPDEs) / (Scalar(nPositive(rowIdx, entryIdx % nPDEs)) != zero ? Scalar(nPositive(rowIdx, entryIdx % nPDEs)) : one));
+          }
+        }
+        for (size_t k = 0; k < (size_t)nPDEs; k++) ConstraintViolationSum(rowIdx, k) = zero;
+      }
+      for (size_t k = 0; k < (size_t)nPDEs; k++) Rsum(rowIdx, k) = zero;
+      for (size_t k = 0; k < (size_t)nPDEs; k++) nPositive(rowIdx, k) = 0;
+    }  // while (checkRow) ...
+  }
+};
+
+template <typename local_matrix_type>
+struct optimalSatisfyConstraintsForScalarPDEsKernel {
+  using Scalar      = typename local_matrix_type::non_const_value_type;
+  using SC          = Scalar;
+  using LO          = typename local_matrix_type::non_const_ordinal_type;
+  using Device      = typename local_matrix_type::device_type;
+  using KAT         = Kokkos::ArithTraits<SC>;
+  const Scalar zero = KAT::zero();
+  const Scalar one  = KAT::one();
+  LO nPDEs;
+  local_matrix_type localP;
+  Kokkos::View<SC**, Device> origSorted;
+  Kokkos::View<SC**, Device> fixedSorted;
+  Kokkos::View<LO**, Device> inds;
+
+  optimalSatisfyConstraintsForScalarPDEsKernel(LO nPDEs_, LO maxRowEntries_, local_matrix_type localP_)
+    : nPDEs(nPDEs_)
+    , localP(localP_) {
+    origSorted  = Kokkos::View<SC**, Device>("origSorted", localP_.numRows(), maxRowEntries_);
+    fixedSorted = Kokkos::View<SC**, Device>("fixedSorted", localP_.numRows(), maxRowEntries_);
+    inds        = Kokkos::View<LO**, Device>("inds", localP_.numRows(), maxRowEntries_);
+  }
+
+  KOKKOS_INLINE_FUNCTION
+  void operator()(const size_t rowIdx) const {
+    auto rowPtr = localP.graph.row_map;
+    auto values = localP.values;
+
+    auto nnz = rowPtr(rowIdx + 1) - rowPtr(rowIdx);
+
+    if (nnz != 0) {
+      Scalar rsumTarget = zero;
+      for (auto entryIdx = rowPtr(rowIdx); entryIdx < rowPtr(rowIdx + 1); entryIdx++) rsumTarget += values(entryIdx);
+
+      {
+        SC aBigNumber;
+        SC rowSumDeviation, temp, delta;
+        SC closestToLeftBoundDist, closestToRghtBoundDist;
+        LO closestToLeftBound, closestToRghtBound;
+        bool flipped;
+
+        SC leftBound = zero;
+        SC rghtBound = one;
+        if ((KAT::real(rsumTarget) >= KAT::real(leftBound * (static_cast<SC>(nnz)))) &&
+            (KAT::real(rsumTarget) <= KAT::real(rghtBound * (static_cast<SC>(nnz))))) {  // has Feasible solution
+
+          flipped = false;
+          // compute aBigNumber to handle some corner cases where we need
+          // something large so that an if statement will be false
+          aBigNumber = KAT::zero();
+          for (auto entryIdx = rowPtr(rowIdx); entryIdx < rowPtr(rowIdx + 1); entryIdx++) {
+            if (KAT::magnitude(values(entryIdx)) > KAT::magnitude(aBigNumber))
+              aBigNumber = KAT::magnitude(values(entryIdx));
+          }
+          aBigNumber = aBigNumber + (KAT::magnitude(leftBound) + KAT::magnitude(rghtBound)) * (100 * one);
+
+          LO ind = 0;
+          for (auto entryIdx = rowPtr(rowIdx); entryIdx < rowPtr(rowIdx + 1); entryIdx++) {
+            origSorted(rowIdx, ind) = values(entryIdx);
+            inds(rowIdx, ind)       = ind;
+            ind++;
+          }
+
+          auto sortVals = Kokkos::subview(origSorted, rowIdx, Kokkos::ALL());
+          auto sortInds = Kokkos::subview(inds, rowIdx, Kokkos::make_pair(0, ind));
+          // Need permutation indices to sort row values from smallest to largest,
+          // and unsort once row constraints are applied.
+          // serial insertion sort workaround from https://github.com/kokkos/kokkos/issues/645
+          for (LO i = 1; i < static_cast<LO>(nnz); ++i) {
+            ind  = sortInds(i);
+            LO j = i;
+
+            if (KAT::real(sortVals(sortInds(i))) < KAT::real(sortVals(sortInds(0)))) {
+              for (; j > 0; --j) sortInds(j) = sortInds(j - 1);
+
+              sortInds(0) = ind;
+            } else {
+              for (; KAT::real(sortVals(ind)) < KAT::real(sortVals(sortInds(j - 1))); --j) sortInds(j) = sortInds(j - 1);
+
+              sortInds(j) = ind;
+            }
+          }
+
+          for (LO i = 0; i < static_cast<LO>(nnz); i++) origSorted(rowIdx, i) = values(rowPtr(rowIdx) + inds(rowIdx, i));  // values is no longer used
+          // find entry in origSorted just to the right of the leftBound
+          closestToLeftBound = 0;
+          while ((closestToLeftBound < static_cast<LO>(nnz)) && (KAT::real(origSorted(rowIdx, closestToLeftBound)) <= KAT::real(leftBound))) closestToLeftBound++;
+
+          // find entry in origSorted just to the right of the rghtBound
+          closestToRghtBound = closestToLeftBound;
+          while ((closestToRghtBound < static_cast<LO>(nnz)) && (KAT::real(origSorted(rowIdx, closestToRghtBound)) <= KAT::real(rghtBound))) closestToRghtBound++;
+
+          // compute distance between closestToLeftBound and the left bound and the
+          // distance between closestToRghtBound and the right bound.
+
+          closestToLeftBoundDist = origSorted(rowIdx, closestToLeftBound) - leftBound;
+          if (closestToRghtBound == static_cast<LO>(nnz))
+            closestToRghtBoundDist = aBigNumber;
+          else
+            closestToRghtBoundDist = origSorted(rowIdx, closestToRghtBound) - rghtBound;
+
+          // compute how far the rowSum is off from the target row sum taking into account
+          // numbers that have been shifted to satisfy bound constraint
+
+          rowSumDeviation = leftBound * (static_cast<SC>(closestToLeftBound)) + (static_cast<SC>(nnz - closestToRghtBound)) * rghtBound - rsumTarget;
+          for (LO i = closestToLeftBound; i < closestToRghtBound; i++) rowSumDeviation += origSorted(rowIdx, i);
+
+          // the code that follow after this if statement assumes that rowSumDeviation is positive. If this
+          // is not the case, flip the signs of everything so that rowSumDeviation is now positive.
+          // Later we will flip the data back to its original form.
+          if (KAT::real(rowSumDeviation) < KAT::real(KAT::zero())) {
+            flipped   = true;
+            temp      = leftBound;
+            leftBound = -rghtBound;
+            rghtBound = temp;
+
+            /* flip sign of origSorted and reverse ordering so that the negative version is sorted */
+
+            // TODO: the following is bad for GPU performance. Switch to bit shifting if brave.
+            if ((nnz % 2) == 1) origSorted(rowIdx, (nnz / 2)) = -origSorted(rowIdx, (nnz / 2));
+            for (LO i = 0; i < static_cast<LO>(nnz / 2); i++) {
+              temp                            = origSorted(rowIdx, i);
+              origSorted(rowIdx, i)           = -origSorted(rowIdx, nnz - 1 - i);
+              origSorted(rowIdx, nnz - i - 1) = -temp;
+            }
+
+            /* reverse bounds */
+
+            LO itemp               = closestToLeftBound;
+            closestToLeftBound     = nnz - closestToRghtBound;
+            closestToRghtBound     = nnz - itemp;
+            closestToLeftBoundDist = origSorted(rowIdx, closestToLeftBound) - leftBound;
+            if (closestToRghtBound == static_cast<LO>(nnz))
+              closestToRghtBoundDist = aBigNumber;
+            else
+              closestToRghtBoundDist = origSorted(rowIdx, closestToRghtBound) - rghtBound;
+
+            rowSumDeviation = -rowSumDeviation;
+          }
+
+          // initial fixedSorted so bounds are satisfied and interiors correspond to origSorted
+
+          for (LO i = 0; i < closestToLeftBound; i++) fixedSorted(rowIdx, i) = leftBound;
+          for (LO i = closestToLeftBound; i < closestToRghtBound; i++) fixedSorted(rowIdx, i) = origSorted(rowIdx, i);
+          for (LO i = closestToRghtBound; i < static_cast<LO>(nnz); i++) fixedSorted(rowIdx, i) = rghtBound;
+
+          while ((KAT::magnitude(rowSumDeviation) > KAT::magnitude((one * 1.e-10) * rsumTarget))) {  // && ( (closestToLeftBound < nEntries ) || (closestToRghtBound < nEntries))) {
+            if (closestToRghtBound != closestToLeftBound)
+              delta = rowSumDeviation / static_cast<SC>(closestToRghtBound - closestToLeftBound);
+            else
+              delta = aBigNumber;
+
+            if (KAT::magnitude(closestToLeftBoundDist) <= KAT::magnitude(closestToRghtBoundDist)) {
+              if (KAT::magnitude(delta) <= KAT::magnitude(closestToLeftBoundDist)) {
+                rowSumDeviation = zero;
+                for (LO i = closestToLeftBound; i < closestToRghtBound; i++) fixedSorted(rowIdx, i) = origSorted(rowIdx, i) - delta;
+              } else {
+                rowSumDeviation                         = rowSumDeviation - closestToLeftBoundDist;
+                fixedSorted(rowIdx, closestToLeftBound) = leftBound;
+                closestToLeftBound++;
+                if (closestToLeftBound < static_cast<LO>(nnz))
+                  closestToLeftBoundDist = origSorted(rowIdx, closestToLeftBound) - leftBound;
+                else
+                  closestToLeftBoundDist = aBigNumber;
+              }
+            } else {
+              if (KAT::magnitude(delta) <= KAT::magnitude(closestToRghtBoundDist)) {
+                rowSumDeviation = 0;
+                for (LO i = closestToLeftBound; i < closestToRghtBound; i++) fixedSorted(rowIdx, i) = origSorted(rowIdx, i) - delta;
+              } else {
+                rowSumDeviation = rowSumDeviation + closestToRghtBoundDist;
+                //        if (closestToRghtBound < nEntries) {
+                fixedSorted(rowIdx, closestToRghtBound) = origSorted(rowIdx, closestToRghtBound);
+                closestToRghtBound++;
+                //       }
+                if (closestToRghtBound >= static_cast<LO>(nnz))
+                  closestToRghtBoundDist = aBigNumber;
+                else
+                  closestToRghtBoundDist = origSorted(rowIdx, closestToRghtBound) - rghtBound;
+              }
+            }
+          }
+
+          auto rowStart = rowPtr(rowIdx);
+          if (flipped) {
+            /* flip sign of fixedSorted and reverse ordering so that the positve version is sorted */
+
+            if ((nnz % 2) == 1) fixedSorted(rowIdx, (nnz / 2)) = -fixedSorted(rowIdx, (nnz / 2));
+            for (LO i = 0; i < static_cast<LO>(nnz / 2); i++) {
+              temp                             = fixedSorted(rowIdx, i);
+              fixedSorted(rowIdx, i)           = -fixedSorted(rowIdx, nnz - 1 - i);
+              fixedSorted(rowIdx, nnz - i - 1) = -temp;
+            }
+          }
+          // unsort and update row values with new values
+          for (LO i = 0; i < static_cast<LO>(nnz); i++) values(rowStart + inds(rowIdx, i)) = fixedSorted(rowIdx, i);
+
+        } else {  // row does not have feasible solution to match constraint
+          // just set all entries to the same value giving a row sum of 1
+          for (auto entryIdx = rowPtr(rowIdx); entryIdx < rowPtr(rowIdx + 1); entryIdx++) values(entryIdx) = one / (static_cast<SC>(nnz));
+        }
+      }
+    }
+  }
+};
+
+template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
+void SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::SatisfyPConstraints(const RCP<Matrix> A, RCP<Matrix>& P) const {
+  using Device = typename Matrix::local_matrix_type::device_type;
+  LO nPDEs     = A->GetFixedBlockSize();
+
+  using local_mat_type = typename Matrix::local_matrix_type;
+  constraintKernel<local_mat_type> myKernel(nPDEs, P->getLocalMatrixDevice());
+  Kokkos::parallel_for("enforce constraint", Kokkos::RangePolicy<typename Device::execution_space>(0, P->getRowMap()->getLocalNumElements()),
+                       myKernel);
+
+}  // SatsifyPConstraints()
+
+template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
+void SaPFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::optimalSatisfyPConstraintsForScalarPDEs(RCP<Matrix>& P) const {
+  using Device = typename Matrix::local_matrix_type::device_type;
+  LO nPDEs     = 1;  // A->GetFixedBlockSize();
+
+  using local_mat_type = typename Matrix::local_matrix_type;
+  optimalSatisfyConstraintsForScalarPDEsKernel<local_mat_type> myKernel(nPDEs, P->getLocalMaxNumRowEntries(), P->getLocalMatrixDevice());
+  Kokkos::parallel_for("enforce constraint", Kokkos::RangePolicy<typename Device::execution_space>(0, P->getLocalNumRows()),
+                       myKernel);
+
+}  // SatsifyPConstraints()
 
 }  // namespace MueLu
