+++ packages/muelu/src/Utils/ExplicitInstantiation/ETI_SC_LO_GO_NO_classes.cmake
@@ -95,7 +95,7 @@ APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::RepartitionHeuristicFactory-.?if
 APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::ReplicatePFactory )
 APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::RigidBodyModeFactory )
 APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::SaPFactory )
-APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::SaPFactory_kokkos )
+APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::SaPFactory_kokkos-.?if.defined[HAVE_MUELU_DEPRECATED_CODE] )
 APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::ScaledNullspaceFactory )
 APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::SchurComplementFactory )
 APPEND_SET(MUELU_SC_LO_GO_NO_ETI_CLASSES MueLu::SegregatedAFactory )
