+++ packages/muelu/src/Utils/ForwardDeclaration/MueLu_SaPFactory_kokkos_fwd.hpp
@@ -46,6 +46,9 @@
 #ifndef MUELU_SAPFACTORY_KOKKOS_FWD_HPP
 #define MUELU_SAPFACTORY_KOKKOS_FWD_HPP
 
+#include "MueLu_ConfigDefs.hpp"
+#if defined(HAVE_MUELU_DEPRECATED_CODE)
+
 namespace MueLu {
 template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
 class SaPFactory_kokkos;
@@ -53,6 +56,8 @@ class SaPFactory_kokkos;
 
 #ifndef MUELU_SAPFACTORY_KOKKOS_SHORT
 #define MUELU_SAPFACTORY_KOKKOS_SHORT
+#endif
+
 #endif
 
 #endif  // MUELU_SAPFACTORY_KOKKOS_FWD_HPP
