+++ packages/muelu/test/structured/structured_interp_sa_kokkos.xml
@@ -44,7 +44,7 @@
     </ParameterList>
 
     <ParameterList name="mySaPFact">
-      <Parameter name="factory"                             type="string" value="SaPFactory_kokkos"/>
+      <Parameter name="factory"                             type="string" value="SaPFactory"/>
       <Parameter name="P"                                   type="string" value="myProlongatorFact"/>
       <Parameter name="sa: damping factor"                  type="double" value="1.333"/>
     </ParameterList>
